Move spec helpers from spec_helpers/ to helpers/
