+++ queue/spec/backends/sqs/retry_logic_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-require 'spec_helpers/sqs/sqs_test_message'
+require 'helpers/sqs/sqs_test_message'
 
 include TxghQueue
 include TxghQueue::Backends
