+++ queue/spec/helpers/nil_logger.rb
