+++ queue/spec/helpers/sqs/sqs_test_message.rb
