+++ queue/spec/helpers/test_backend.rb
@@ -2,24 +2,41 @@ module TxghQueue
   class TestBackend
     class << self
       def producer_for(event, logger = Txgh::TxLogger.logger)
-        TestProducer.new(event, logger)
+        producers[event] ||= TestProducer.new(event, logger)
       end
 
       def consumer_for(event, logger = Txgh::TxLogger.logger)
-        TestConsumer.new(event, logger)
+        consumers[event] ||= TestConsumer.new(event, logger)
+      end
+
+      def reset!
+        @producers = nil
+        @consumers = nil
+      end
+
+      private
+
+      def producers
+        @producers ||= {}
+      end
+
+      def consumers
+        @consumers ||= {}
       end
     end
   end
 
   class TestProducer
-    attr_reader :queue_names, :logger
+    attr_reader :queue_names, :logger, :enqueued_jobs
 
     def initialize(event, logger)
       @event = event
       @logger = logger
+      @enqueued_jobs = []
     end
 
     def enqueue(payload, options = {})
+      enqueued_jobs << { payload: payload, options: options }
     end
   end
